# web/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CKA Labs FastAPI backend",
    "version": "0.1.0",
    "description": "OpenAPI snapshot of the FastAPI surface consumed by the Vue frontend (FE-004). The live spec is preferred (POST http://<host>:3000/openapi.json); this committed snapshot keeps `gen:api` fully offline. Shapes mirror web/server.py Pydantic models and route return values. NOTE: this snapshot is hand-authored because the platform host accepted the TCP connection but stalled before sending the response body (see .agents/FE-004-implementer-2/EVIDENCE.md)."
  },
  "servers": [{ "url": "/" }],
...
  "paths": {
...
    "/api/presets/select": {
      "post": {
        "operationId": "selectPreset",
        "tags": ["presets"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PresetSelectRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preset selected",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PresetSelectResponse" }
              }
            }
          }
        }
      }
    },
    "/api/start": {
      "post": {
        "operationId": "startSession",
        "tags": ["session"],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/StartRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Newly started session",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "429": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/action/next": {
      "post": {
        "operationId": "actionNext",
        "tags": ["actions"],
        "responses": {
          "200": {
            "description": "Session advanced to the next task",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/action/prev": {
      "post": {
        "operationId": "actionPrev",
        "tags": ["actions"],
        "responses": {
          "200": {
            "description": "Session moved to the previous task",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/action/jump": {
      "post": {
        "operationId": "actionJump",
        "tags": ["actions"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/JumpRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session jumped to the requested task",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/action/flag": {
      "post": {
        "operationId": "actionFlag",
        "tags": ["actions"],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FlagRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Flag toggled",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FlagResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/action/retry": {
      "post": {
        "operationId": "actionRetry",
        "tags": ["actions"],
        "responses": {
          "200": {
            "description": "Current task re-initialised",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StatusMessageResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/action/submit": {
      "post": {
        "operationId": "actionSubmit",
        "tags": ["actions"],
        "responses": {
          "200": {
            "description": "Final scorecard report",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SubmitResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "Error": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorDetail" }
          }
        }
      }
    },
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "required": ["detail"],
        "properties": {
          "detail": { "type": "string" }
        }
      },
      "StatusMessageResponse": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": { "type": "string" },
          "message": { "type": "string" }
        }
      },
      "PresetLockedInfo": {
        "type": "object",
        "required": [
          "filename",
          "name",
          "description",
          "task_count",
          "time_limit_minutes",
          "pass_threshold_percent"
        ],
        "properties": {
          "filename": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "task_count": { "type": "integer" },
          "time_limit_minutes": { "type": ["integer", "null"] },
          "pass_threshold_percent": { "type": "number" }
        }
      },
      "TaskScoreData": {
        "type": "object",
        "required": ["passed", "score", "max_score", "message"],
        "properties": {
          "passed": { "type": "boolean" },
          "score": { "type": "integer" },
          "max_score": { "type": "integer" },
          "message": { "type": "string" }
        }
      },
      "TaskData": {
        "type": "object",
        "required": [
          "task_num",
          "id",
          "title",
          "domain",
          "difficulty",
          "points",
          "target_context",
          "namespace",
          "description",
          "cluster_scoped",
          "tags",
          "is_flagged"
        ],
        "properties": {
          "task_num": { "type": "integer" },
          "id": { "type": "string" },
          "title": { "type": "string" },
          "domain": { "type": "string" },
          "difficulty": { "type": "string", "enum": ["easy", "medium", "hard", "crazy"] },
          "points": { "type": "integer" },
          "target_context": { "type": "string" },
          "namespace": { "type": "string" },
          "description": { "type": "string" },
          "cluster_scoped": { "type": "boolean" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "is_flagged": { "type": "boolean" },
          "score_data": {
            "oneOf": [{ "$ref": "#/components/schemas/TaskScoreData" }, { "type": "null" }]
          }
        }
      },
      "SessionActive": {
        "type": "object",
        "required": [
          "active",
          "session_id",
          "name",
          "mode",
          "status",
          "current_index",
          "total_tasks",
          "time_remaining_seconds",
          "created_at",
          "server_timestamp",
          "flagged_ids",
          "is_admin"
        ],
        "properties": {
          "active": { "type": "boolean", "const": true },
          "session_id": { "type": "string" },
          "candidate_token": { "type": ["string", "null"] },
          "name": { "type": "string" },
          "mode": { "type": "string" },
          "status": { "type": "string" },
          "current_index": { "type": "integer" },
          "total_tasks": { "type": "integer" },
          "time_limit_minutes": { "type": ["integer", "null"] },
          "time_remaining_seconds": { "type": ["number", "null"] },
          "created_at": { "type": "string" },
          "start_timestamp": { "type": ["number", "null"] },
          "end_timestamp": { "type": ["number", "null"] },
          "server_timestamp": { "type": "number" },
          "flagged_ids": { "type": "array", "items": { "type": "string" } },
          "current_task": {
            "oneOf": [{ "$ref": "#/components/schemas/TaskData" }, { "type": "null" }]
          },
          "terminal_port": { "type": "string" },
          "novnc_port": { "type": "string" },
          "is_admin": { "type": "boolean" }
        }
      },
      "SessionInactive": {
        "type": "object",
        "required": ["active", "session", "locked_preset", "is_admin"],
        "properties": {
          "active": { "type": "boolean", "const": false },
          "session": { "type": "null" },
          "locked_preset": { "$ref": "#/components/schemas/PresetLockedInfo" },
          "is_admin": { "type": "boolean" }
        }
      },
      "SessionInvited": {
        "type": "object",
        "required": ["active", "invited", "preset", "token", "locked_preset", "is_admin"],
        "properties": {
          "active": { "type": "boolean", "const": false },
          "invited": { "type": "boolean", "const": true },
          "preset": { "type": "string" },
          "token": { "type": "string" },
          "locked_preset": { "$ref": "#/components/schemas/PresetLockedInfo" },
          "is_admin": { "type": "boolean" }
        }
      },
      "SessionResponse": {
        "oneOf": [
          { "$ref": "#/components/schemas/SessionActive" },
          { "$ref": "#/components/schemas/SessionInactive" },
          { "$ref": "#/components/schemas/SessionInvited" }
        ]
      },
...
      "PresetSelectRequest": {
        "type": "object",
        "required": ["preset"],
        "properties": {
          "preset": { "type": "string" }
        }
      },
      "PresetSelectResponse": {
        "type": "object",
        "required": ["status", "preset", "info"],
        "properties": {
          "status": { "type": "string" },
          "preset": { "type": "string" },
          "info": { "$ref": "#/components/schemas/PresetLockedInfo" }
        }
      },
      "StartRequest": {
        "type": "object",
        "properties": {
          "preset": { "type": ["string", "null"] },
          "all_questions": { "type": ["boolean", "null"] },
          "question_ids": {
            "oneOf": [{ "type": "array", "items": { "type": "string" } }, { "type": "null" }]
          },
          "candidate_token": { "type": ["string", "null"] }
        }
      },
...
      "JumpRequest": {
        "type": "object",
        "required": ["task_num"],
        "properties": {
          "task_num": { "type": "integer" }
        }
      },
      "FlagRequest": {
        "type": "object",
        "properties": {
          "task_num": { "type": ["integer", "null"] }
        }
      },
      "FlagResponse": {
        "type": "object",
        "required": ["id", "task_num", "is_flagged", "flagged_ids"],
        "properties": {
          "id": { "type": "string" },
          "task_num": { "type": "integer" },
          "is_flagged": { "type": "boolean" },
          "flagged_ids": { "type": "array", "items": { "type": "string" } }
        }
      },
      "ScorecardRow": {
        "type": "object",
        "required": [
          "task_num",
          "id",
          "title",
          "domain",
          "difficulty",
          "context",
          "score",
          "max_score",
          "passed",
          "message"
        ],
        "properties": {
          "task_num": { "type": "integer" },
          "id": { "type": "string" },
          "title": { "type": "string" },
          "domain": { "type": "string" },
          "difficulty": { "type": "string" },
          "context": { "type": "string" },
          "score": { "type": "integer" },
          "max_score": { "type": "integer" },
          "passed": { "type": "boolean" },
          "message": { "type": "string" }
        }
      },
      "SubmitResponse": {
        "type": "object",
        "required": [
          "scorecard",
          "total_earned",
          "total_possible",
          "percentage",
          "passed",
          "threshold",
          "session_id",
          "exam_name",
          "submitted_at"
        ],
        "properties": {
          "scorecard": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ScorecardRow" }
          },
          "total_earned": { "type": "integer" },
          "total_possible": { "type": "integer" },
          "percentage": { "type": "number" },
          "passed": { "type": "boolean" },
          "threshold": { "type": "number" },
          "session_id": { "type": "string" },
          "exam_name": { "type": "string" },
          "submitted_at": { "type": "string" },
          "report_file": { "type": "string" }
        }
      },
...
    }
  }
}